Updated Maven Assembly Plugin version to 3.0.0.
This resolves #5.

Also fixed a warning emitted by the Maven Assembly Plugin that was
caused by the use of both zip and dir formats in the assembly
configuration. The solution was to simply comment out the dir format.
Since it can be useful to have the build output to a dir, a comment in
the assembly configuration file explains how to enable it.