Remove server-sdk.maven-version
This replaces a workaround for PingData servers rejecting SNAPSHOT
extensions with one that does not require extension developers to
maintain a separate server-sdk.maven-version. The new workaround uses
the same maven-install-plugin mechanism that strips the '-EA' segment of
the versioning from the extension metadata. It now strips both '-EA' and
'-SNAPSHOT'.