Fix Java 11 compilation issues
This modifies the server-sdk-archetype to fix a couple of issues when
compiling on Java 11. First, it specifies the javadoc executable,
working around a problem where the maven-javadoc-plugin relies on
JAVA_HOME to find the executable (MJAVADOC-595). Second, it fixes a
problem with javadoc and classes not belonging to any explicit Java
modules by explicitly configuring javadoc to expect Java 8 source.